pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "domain-learning-agent"
version = "0.1.0"
description = "PhD-level AI research domain learning system with 12 agent skills"
readme = "README.md"
requires-python = ">=3.14"
license = {text = "MIT"}
authors = [
    {name = "Kevin"},
]

dependencies = [
    "pydantic>=2.0",
    "httpx>=0.27",
    "networkx>=3.0",
    "jinja2>=3.1",
    "rich>=13.0",
    "typer>=0.15",
    "genanki>=0.13",
    "nbformat>=5.10",
    "matplotlib>=3.9",
    "anthropic>=0.42",
    "pyyaml>=6.0",
    "aiohttp>=3.11",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-dotenv>=1.0",
    "sse-starlette>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "respx>=0.22",
]
pdf = [
    "pypandoc>=1.14",
]

[project.scripts]
newlearner = "src.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests that require API access",
    "slow: marks tests that take a long time to run",
]
